Moved sqlite resources to cache modulke
